pyproject.toml
[project]
name = "targetfit"
version = "0.1.0"
description = "Local-first job matching tool — scrape career pages, embed with Ollama, rank against your CV."
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4",
    "click",
    "duckdb",
    "numpy",
    "ollama",
    "playwright",
    "pyyaml",
    "requests",
    "rich",
    "scrapegraphai",
]

[project.scripts]
targetfit = "targetfit.cli:cli"

[project.optional-dependencies]
dev = ["pytest", "ruff"]

[tool.setuptools.packages.find]
include = ["targetfit*"]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"
